Fixed refactoring regressions
