+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/testing/adapter/TestResultViewHolder.kt
@@ -11,7 +11,7 @@ import com.ftinc.kit.arch.util.bindView
 import com.r0adkll.deckbuilder.GlideApp
 import com.r0adkll.deckbuilder.R
 import com.r0adkll.deckbuilder.arch.ui.widgets.TestResultProgressView
-import com.r0adkll.deckbuilder.util.extensions.ImageType
+import com.r0adkll.deckbuilder.util.glide.ImageType
 import com.r0adkll.deckbuilder.util.extensions.loadPokemonCard
 
 @Suppress("MagicNumber")
